+++ Assets/Scripts/MB/ResourceReservation.cs
@@ -4,5 +4,5 @@ using System.Collections;
 public class ResourceReservation : Reservation {
     public GameObject source;
     public string resourceTag;
-    public float amount;
+    public decimal amount;
 }
